+++ packages/app/hivecommand-native/src/App.tsx
@@ -13,41 +13,28 @@ import { BaseDirectory } from '@tauri-apps/api/fs';
 import { Controller } from './views/controller';
 
 import { Command } from '@tauri-apps/api/shell';
+import axios from 'axios';
 
 const cmd = Command.sidecar('binaries/sidecar')
 
 function App() {
 
-  const { authState } = useContext(DataContext)
+  const { authState, setAuthState, updateAuthState } = useContext(DataContext)
 
   useEffect(() => {
     cmd.execute().then((proc) => {
       console.log("Sidecar running");
 
-      
+      axios.get(`http://localhost:${8484}/setup`).then((data) => {
+        if(data.config){
+          setAuthState?.({...authState, configProvided: true} )
+        }
+      })
+
     });
   }, [])
   
-  
-  useEffect(() => {
-    console.log({ BaseDirectory: BaseDirectory.App })
-    
-    
-    // createDir('conf', {dir: BaseDirectory.App, recursive: true}).then(() => {
-    //   writeTextFile({path: CONF_FILE, contents: '{}'}, {dir: BaseDirectory.App}).then(() => {
-    //     console.log("ASFD")
 
-    //     readTextFile(CONF_FILE, {dir: BaseDirectory.App}).then((confText: any) => {
-
-    //       if(confText){
-    //         setConf(JSON.parse(confText))
-    //       }
-
-    //     })
-    //   })
-    // })
-
-  }, [])
 
   const renderView = () => {
     if (!authState?.isAuthed()) {