+++ packages/communication/opc-mqtt-auth/src/index.ts
@@ -1,0 +1,47 @@
+import {Router} from 'express';
+
+export const MQTTAuth = (
+    getUser: ( username: string, password: string ) => Promise<boolean>,
+    isValidResource: (username: string, name: string) => Promise<boolean> 
+) => {
+    const router = Router();
+
+    router.get('/auth/user', async (req, res) => {
+        console.log(req.query);
+        const userQuery : {username: string, password: string} = req.query as any;
+
+        try{
+            const isAuthed = await getUser(userQuery.username, userQuery.password);
+            if(isAuthed){
+                res.send('allow device hive-command');
+            }else{
+                res.send('deny')
+            }
+        }catch(e){
+            res.send('deny')
+        }
+    });
+
+    router.get('/auth/vhost', (req, res) => {
+        res.send("allow");
+    });
+
+    router.get('/auth/resource', async (req, res) => {
+        const { username, name } = req.query as any;
+        const isValid = await isValidResource(username, name);
+
+        if((isValid && (req.query.permission === 'configure' || req.query.permission === "permission") || req.query.permission === "write") || req.query.name === 'amq.default'){
+            return res.send("allow");
+        }else{
+            console.log("Resources failed ", req.query)
+        }
+
+        return res.send('deny');
+    });
+
+    router.get('/auth/topic', (req, res) => {
+        res.send('allow');
+    });
+
+    return router;
+}