+++ packages/infrastructure/src/mqtt/deployment.ts
@@ -2,12 +2,29 @@ import {Provider} from '@pulumi/kubernetes'
 import * as k8s from '@pulumi/kubernetes'
 import { Config } from '@pulumi/pulumi';
 
-export const RabbitMQDeployment = async (provider: Provider, appName: string, storageClaim: k8s.core.v1.PersistentVolumeClaim, ns: k8s.core.v1.Namespace) => {
+export const RabbitMQDeployment = async (provider: Provider, appName: string, storageClaim: k8s.core.v1.PersistentVolumeClaim, authApi: string, ns: k8s.core.v1.Namespace) => {
 
     const config = new Config();
     const suffix = config.require('suffix');
 
     const appLabels = { appClass: appName };
+
+    const configMap = new k8s.core.v1.ConfigMap(`${appName}-config`, {
+        data: {
+            'rabbitmq.conf': `log.console = true
+log.console.level = debug
+
+auth_backends.1 = rabbit_auth_backend_http
+
+auth_http.user_path = ${authApi}/auth/user
+auth_http.vhost_path = ${authApi}/auth/vhost
+auth_http.resource_path = ${authApi}/auth/resource
+auth_http.topic_path = ${authApi}/auth/topic`,
+            'enabled_plugins': '[rabbitmq_prometheus,rabbitmq_auth_backend_http].'
+        }
+    }, {
+        provider
+    })
 
     const deployment = new k8s.apps.v1.Deployment(`${appName}-dep`, {
         metadata: { 
@@ -32,6 +49,10 @@ export const RabbitMQDeployment = async (provider: Provider, appName: string, st
                             {
                                 name: 'persistence',
                                 mountPath: `/var/lib/rabbitmq/mnesia/`
+                            },
+                            {
+                                name: 'config',
+                                mountPath: '/etc/rabbitmq'
                             }
                         ],
                         // env: [
@@ -48,6 +69,12 @@ export const RabbitMQDeployment = async (provider: Provider, appName: string, st
                             name: 'persistence',
                             persistentVolumeClaim: {
                                 claimName: storageClaim.metadata.name
+                            }
+                        },
+                        {
+                            name: 'config',
+                            configMap: {
+                                name: configMap.metadata.name
                             }
                         }
                     ]